+++ CHANGELOG.md
@@ -5,8 +5,22 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Emit event on producer mutation.
+- Emit event on streambinding mutation.
+
+### Bugfix
+- Fixed issues with referential integrity on binding entities.
+
+## [0.10.5] 2020-03-24
 ### Changed
 - Removed auto configuration of default handlers.
+
+### Added
+- A listener for streamBinding events
+- A Kafka handler for streamBinding events
+- A listener for producer events
+- A Kafka handler for producer events
 
 ## [0.10.4] 2020-03-20
 ### Added